% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSchedule.R
\name{checkSchedule}
\alias{checkSchedule}
\title{checkSchedule}
\usage{
checkSchedule(schedule, minute_range)
}
\arguments{
\item{schedule}{A \code{\link{schedule}} object.}

\item{minute_range}{The range in minutes to allow for a different, and still post. Defaults to 3.}
}
\description{
Check if any of the parsed schedules match the current time.
}
\examples{
now <- crontweets:::nowFormatted()
data(schedule)
scheduleCheck(schedule = schedule, minute_range=5)
}
