% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkScheduleAndPost.R
\name{checkScheduleAndPost}
\alias{checkScheduleAndPost}
\title{checkScheduleAndPost}
\usage{
checkScheduleAndPost(schedule, tweet_db = NULL)
}
\arguments{
\item{schedule}{A \code{\link{schedule}} object.}

\item{tweet_db}{Either an in-memory \code{\link{tweet_db}} object, or a file path to one. This is
a convenience function which sits inside other functions just to parse path vs object.}
}
\description{
A top level function usually called by cron and a command-line-interface.
}
\examples{
data(tweet_db)
data(schedule)
checkScheduleAndPost(schedule, tweet_db)
}
