% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossCheckScheduleTweetDB.R
\name{crossCheckScheduleTweetDB}
\alias{crossCheckScheduleTweetDB}
\title{crossCheckScheduleTweetDB}
\usage{
crossCheckScheduleTweetDB(schedule, tweet_db)
}
\arguments{
\item{schedule}{A \code{\link{schedule}} object.}
}
\description{
Cross-checks a \code{\link{schedule}} against a \code{\link{tweetdb}} to ensure no missing entries in either object.
}
\examples{
data(tweetdb)
data(schedule)
crossCheckScheduleTweetDB(schedule, tweet_db)
}
