% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetByCategory.R
\name{getTweetByCategory}
\alias{getTweetByCategory}
\title{getTweetByCategory}
\usage{
getTweetByCategory(category, n = 1, tweet_db)
}
\arguments{
\item{category}{A category to filter the resulting tweets by.  \code{NULL} by default.}

\item{n}{How many tweets do you want? Defaults to 1, but if you use 'all', it will pull all.}

\item{tweet_db}{Either a tweet_db object, or a filepath to such an object.}
}
\description{
Get a random set of tweets from the \code{\link{tweet_db}}.
}
\examples{
data(tweet_db)
getTweetByCategory(category="animals", tweet_db=tweet_db)
}
