% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetMatches.R
\name{getTweetMatch}
\alias{getTweetMatch}
\title{getTweetByCategory}
\usage{
getTweetMatch(schedule_item, tweet_db)
}
\arguments{
\item{schedule_item}{A single row of a \code{\link{schedule}}.}

\item{tweet_db}{Either an in-memory \code{\link{tweet_db}} object, or a file path to one. This is
a convenience function which sits inside other functions just to parse path vs object.}
}
\description{
Get a random set of tweets from the \code{\link{tweet_db}}.
}
\examples{
data(tweet_db)
getTweetByCategory(category="animals", tweet_db=tweet_db)
}
