% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTweetDB.R
\name{loadTweetDB}
\alias{loadTweetDB}
\title{loadTweetDB}
\usage{
loadTweetDB(tweet_db)
}
\arguments{
\item{tweet_db}{Either an in-memory \code{\link{tweet_db}} object, or a file path to one. This is
a convenience function which sits inside other functions just to parse path vs object.}
}
\value{
A tweet_db object
}
\description{
Loads
}
