% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSchedule.R
\name{nowFormatted}
\alias{nowFormatted}
\title{nowFormatted}
\usage{
nowFormatted()
}
\description{
breaks apart time object to check for matches
}
\examples{
nowFormatted()
Returns the current datetime in a list. 
}
