% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSchedule.R
\name{parseSchedule}
\alias{parseSchedule}
\title{parseSchedule}
\usage{
parseSchedule(schedule, now)
}
\arguments{
\item{schedule}{A \code{\link{schedule}} object.}

\item{now}{Optional. A prespecified time-date object if you want to supply one.}
}
\description{
Parse the items of a schedule, replacing wildcards and dealing with timezones.
}
\examples{
now <- crontwit:::nowFormatted()
data(schedule)
parseSchedule(schedule = schedule, now)
}
