% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTweet.R
\name{postTweet}
\alias{postTweet}
\title{postTweet}
\usage{
postTweet(tweetx, ...)
}
\arguments{
\item{tweetx}{A specific tweet as a list object, with items as in \code{\link{tweet_db}}.}

\item{...}{Optional arguments for \code{\link[twitteR]{tweet}}.}
}
\description{
Posts a tweet after parsing the media path.
}
\examples{
\dontrun{
data(tweet_db)
postTweet(as.list(tweet_db[1,]))
}
}
