% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_db.R
\docType{data}
\name{tweet_db}
\alias{tweet_db}
\title{tweet_db}
\arguments{
\item{tweet_text}{The tweet text.}

\item{category}{The category of tweet.}

\item{media_path}{Optional. The path to any media/pictures you want to tweet with it.}

\item{id}{An automatically generated unique identifier in your \code{\link{tweet_db}}, as generated by \code{\link{addNewTweetToDB}}.}
}
\description{
A data.frame holding the relevant data for the tweet database. Built by \code{\link{addNewTweetToDB}}, and used by \code{\link{getTweetByID}} and \code{\link{getTweetByCategory}}.
}
\examples{
tweet_db <- addNewTweetToDB("I like cats"     , category="animals")
tweet_db <- addNewTweetToDB("I like dogs too.", category="animals", tweet_db = tweet_db)
tweet_db <- addNewTweetToDB("I don't like sardines.", 
                            category="food", tweet_db = tweet_db, ask=FALSE)
tweet_db
}
