% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSchedule.R
\name{validateSchedule}
\alias{validateSchedule}
\title{validateSchedule}
\usage{
validateSchedule(schedule)
}
\arguments{
\item{schedule}{A \code{\link{schedule}} object.}
}
\description{
Validate that all entries in a schedule object are ok.
}
\examples{
data(schedule)
schedule$minute[3] <- 123
validateSchedule(schedule)
}
