% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateTweetDB.R
\name{validateTweetDB}
\alias{validateTweetDB}
\title{validateTweetDB}
\usage{
validateTweetDB(tweet_db)
}
\description{
Validate that all entries in a \code{\link{tweetdb}} object are ok.
}
\examples{
data(tweetdb)
validateTweetDB(tweetdb)
}
