% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewTweetToDB.R
\name{addNewTweetToDB}
\alias{addNewTweetToDB}
\title{addNewTweetToDB}
\usage{
addNewTweetToDB(tweet_text, category, media_path = NA, tweet_db = NULL)
}
\arguments{
\item{tweet_text}{The text of the tweet.}

\item{category}{A category to assign the tweet to.}

\item{media_path}{A file-path to a picture or something you want to share.}

\item{tweet_db}{Allows you to pass an in-memory database in to continue building it out.}
}
\description{
Add a new tweet to the \code{\link{tweet_db}}. If inputs are not specified, it will ask in the console for input. Automatically creats unique identifier.
}
\examples{
addNewTweetToDB("TestTweet", category="test")
}
