% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSchedule.R
\name{filterSchedule}
\alias{filterSchedule}
\title{filterSchedule}
\usage{
filterSchedule(schedule_item, now, minute_range = 3)
}
\arguments{
\item{schedule_item}{A single row of a \code{\link{schedule}}.}

\item{now}{A properlay formatted \code{now} object, as produced by \code{\link{nowFormatted}}.}

\item{minute_range}{The range in minutes to allow for a different, and still post. Defaults to 3.}
}
\description{
Pass a boolean about whether or not a schedule_item is to be posted.
}
\examples{
now <- crontwit:::nowFormatted()
data(schedule)
filterSchedule(schedule_item = schedule[1,], now = now)
}
