% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweetMatches.R
\name{getTweetMatch}
\alias{getTweetMatch}
\title{getTweetByCategory}
\usage{
getTweetMatch(schedule_item, tweet_db)
}
\arguments{
\item{tweet_db}{Either a tweet_db object, or a filepath to such an object.}

\item{category}{A category to filter the resulting tweets by.  \code{NULL} by default.}

\item{n}{How many tweets do you want? Defaults to 1, but if you use 'all', it will pull all.}
}
\description{
Get a random set of tweets from the \code{\link{tweet_db}}.
}
\examples{
data(tweet_db)
getTweetByCategory(category="animals", tweet_db=tweet_db)
}
