% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSchedule.R
\name{nowFormatted}
\alias{nowFormatted}
\title{nowFormatted}
\usage{
nowFormatted(now = NULL)
}
\arguments{
\item{now}{Optional. A prespecified time-date object if you want to supply one.}
}
\description{
Returns the current datetime in a list.
}
\examples{
crontwit:::nowFormatted()
}
